+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.5.1] - 2023-06-09
+
+Fixed:
+  - Masthead: clipping behaviour on small screens
+  - Masthead: improve animation
+
 ## [1.5.0] - 2023-06-09
 
 Added: